+++ game/scrabble.py
@@ -12,6 +12,10 @@ class ScrabbleGame:
             self.players.append(Player())
 
         self.current_player = None
+        #self.turn = 0
+
+    #def playing(self):
+        #retur True
 
     def next_turn(self):
         if self.current_player is None:
@@ -21,4 +25,20 @@ class ScrabbleGame:
             self.current_player = self.players[index]
 
     
+    def validate_word(self, word, location, orientation):
+        '''
+        1- Validar que usuario tiene esas letras
+        2- Validar que la palabra entra en el tablero
+        '''
+        self.board.validate_word_inside_board(word, location, orientation)
     
+    def get_words():
+        '''
+        Obtener las posibles palabras que se pueden formar, dada una palabra, ubicacion y orientacion 
+        Preguntar al usuario, por cada una de esas palabras, las que considera reales
+        '''
+    
+    def put_words():
+        '''
+        Modifica el estado del tablero con las palabras consideradas como correctas
+        '''